requirements.txt
anthropic==0.25.0
playwright==1.44.0
playwright-stealth==1.0.6
Pillow>=10.0.0
python-dotenv==1.0.0
